requirements.txt
biopython (1.66)
DendroPy (4.2.0)
numpy (1.10.4)
pip (8.0.2)
py (1.4.31)
pytest (2.8.5)
setuptools (19.6.2)
wheel (0.29.0)